Refactor service validating code
