+++ compiler-plugin-tests/src/test/java/io/ballerina/stdlib/graphql/compiler/ServiceValidationTest.java
@@ -195,7 +195,7 @@ public class ServiceValidationTest {
     public void testInvalidResourceReturnTypes() {
         String packagePath = "invalid_service_4";
         DiagnosticResult diagnosticResult = getDiagnosticResult(packagePath);
-        Assert.assertEquals(diagnosticResult.errorCount(), 8);
+        Assert.assertEquals(diagnosticResult.errorCount(), 11);
         Iterator<Diagnostic> diagnosticIterator = diagnosticResult.errors().iterator();
 
         Diagnostic diagnostic = diagnosticIterator.next();
@@ -211,13 +211,22 @@ public class ServiceValidationTest {
         assertError(diagnostic, CompilationError.INVALID_RETURN_TYPE, 48, 5);
 
         diagnostic = diagnosticIterator.next();
-        assertError(diagnostic, CompilationError.INVALID_RETURN_TYPE, 55, 5);
+        assertError(diagnostic, CompilationError.INVALID_UNION_MEMBER_TYPE, 55, 5);
 
         diagnostic = diagnosticIterator.next();
-        assertError(diagnostic, CompilationError.INVALID_RETURN_TYPE, 61, 5);
+        assertError(diagnostic, CompilationError.INVALID_UNION_MEMBER_TYPE, 55, 5);
 
         diagnostic = diagnosticIterator.next();
-        assertError(diagnostic, CompilationError.INVALID_RETURN_TYPE, 67, 5);
+        assertError(diagnostic, CompilationError.INVALID_UNION_MEMBER_TYPE, 61, 5);
+
+        diagnostic = diagnosticIterator.next();
+        assertError(diagnostic, CompilationError.INVALID_UNION_MEMBER_TYPE, 61, 5);
+
+        diagnostic = diagnosticIterator.next();
+        assertError(diagnostic, CompilationError.INVALID_UNION_MEMBER_TYPE, 67, 5);
+
+        diagnostic = diagnosticIterator.next();
+        assertError(diagnostic, CompilationError.INVALID_UNION_MEMBER_TYPE, 67, 5);
 
         diagnostic = diagnosticIterator.next();
         assertError(diagnostic, CompilationError.INVALID_RETURN_TYPE, 74, 5);
@@ -326,17 +335,20 @@ public class ServiceValidationTest {
     public void testInvalidResourceReturnTypeUnions() {
         String packagePath = "invalid_service_11";
         DiagnosticResult diagnosticResult = getDiagnosticResult(packagePath);
-        Assert.assertEquals(diagnosticResult.errorCount(), 3);
+        Assert.assertEquals(diagnosticResult.errorCount(), 4);
         Iterator<Diagnostic> diagnosticIterator = diagnosticResult.errors().iterator();
 
         Diagnostic diagnostic = diagnosticIterator.next();
-        assertError(diagnostic, CompilationError.INVALID_RETURN_TYPE_MULTIPLE_SERVICES, 20, 5);
+        assertError(diagnostic, CompilationError.INVALID_UNION_MEMBER_TYPE, 20, 5);
 
         diagnostic = diagnosticIterator.next();
-        assertError(diagnostic, CompilationError.INVALID_RETURN_TYPE_MULTIPLE_SERVICES, 26, 5);
+        assertError(diagnostic, CompilationError.INVALID_UNION_MEMBER_TYPE, 20, 5);
 
         diagnostic = diagnosticIterator.next();
-        assertError(diagnostic, CompilationError.INVALID_RETURN_TYPE_MULTIPLE_SERVICES, 32, 5);
+        assertError(diagnostic, CompilationError.INVALID_UNION_MEMBER_TYPE, 26, 5);
+
+        diagnostic = diagnosticIterator.next();
+        assertError(diagnostic, CompilationError.INVALID_UNION_MEMBER_TYPE, 32, 5);
     }
 
     @Test
