+++ compiler-plugin/src/main/java/io/ballerina/stdlib/graphql/compiler/Utils.java
@@ -18,7 +18,8 @@
 
 package io.ballerina.stdlib.graphql.compiler;
 
-import io.ballerina.compiler.api.SemanticModel;
+import io.ballerina.compiler.api.symbols.ClassSymbol;
+import io.ballerina.compiler.api.symbols.IntersectionTypeSymbol;
 import io.ballerina.compiler.api.symbols.MethodSymbol;
 import io.ballerina.compiler.api.symbols.Qualifier;
 import io.ballerina.compiler.api.symbols.Symbol;
@@ -26,10 +27,12 @@ import io.ballerina.compiler.api.symbols.SymbolKind;
 import io.ballerina.compiler.api.symbols.TypeDescKind;
 import io.ballerina.compiler.api.symbols.TypeReferenceTypeSymbol;
 import io.ballerina.compiler.api.symbols.TypeSymbol;
-import io.ballerina.compiler.syntax.tree.FunctionDefinitionNode;
-import io.ballerina.projects.plugins.SyntaxNodeAnalysisContext;
+import io.ballerina.compiler.api.symbols.UnionTypeSymbol;
+import io.ballerina.tools.diagnostics.Location;
 
-import java.util.Optional;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 
 /**
  * Util class for the compiler plugin.
@@ -57,22 +60,6 @@ public final class Utils {
         return PACKAGE_NAME.equals(moduleName) && PACKAGE_ORG.equals(orgName);
     }
 
-    public static MethodSymbol getMethodSymbol(SyntaxNodeAnalysisContext context,
-                                               FunctionDefinitionNode functionDefinitionNode) {
-        MethodSymbol methodSymbol = null;
-        SemanticModel semanticModel = context.semanticModel();
-        Optional<Symbol> symbol = semanticModel.symbol(functionDefinitionNode);
-        if (symbol.isPresent()) {
-            methodSymbol = (MethodSymbol) symbol.get();
-        }
-        return methodSymbol;
-    }
-
-    public static boolean isRemoteFunction(SyntaxNodeAnalysisContext context,
-                                           FunctionDefinitionNode functionDefinitionNode) {
-        return isRemoteFunction(getMethodSymbol(context, functionDefinitionNode));
-    }
-
     public static boolean isRemoteFunction(MethodSymbol methodSymbol) {
         return methodSymbol.qualifiers().contains(Qualifier.REMOTE);
     }
@@ -92,5 +79,54 @@ public final class Utils {
             return false;
         }
         return LISTENER_IDENTIFIER.equals(typeSymbol.getName().get());
+    }
+
+    public static boolean isIgnoreType(TypeSymbol typeSymbol) {
+        return typeSymbol.typeKind() == TypeDescKind.NIL || typeSymbol.typeKind() == TypeDescKind.ERROR;
+    }
+
+    public static List<TypeSymbol> getEffectiveTypes(UnionTypeSymbol unionTypeSymbol) {
+        List<TypeSymbol> effectiveTypes = new ArrayList<>();
+        for (TypeSymbol typeSymbol : unionTypeSymbol.memberTypeDescriptors()) {
+            if (!isIgnoreType(typeSymbol)) {
+                effectiveTypes.add(typeSymbol);
+            }
+        }
+        return effectiveTypes;
+    }
+
+    public static TypeSymbol getEffectiveType(IntersectionTypeSymbol intersectionTypeSymbol, Location location) {
+        List<TypeSymbol> effectiveTypes = new ArrayList<>();
+        for (TypeSymbol typeSymbol : intersectionTypeSymbol.memberTypeDescriptors()) {
+            if (typeSymbol.typeKind() == TypeDescKind.READONLY) {
+                continue;
+            }
+            effectiveTypes.add(typeSymbol);
+        }
+        if (effectiveTypes.size() == 1) {
+            return effectiveTypes.get(0);
+        }
+        return null;
+    }
+
+    public static boolean isPrimitiveType(TypeSymbol returnType) {
+        return returnType.typeKind().isStringType() ||
+                returnType.typeKind() == TypeDescKind.INT ||
+                returnType.typeKind() == TypeDescKind.FLOAT ||
+                returnType.typeKind() == TypeDescKind.BOOLEAN ||
+                returnType.typeKind() == TypeDescKind.DECIMAL;
+    }
+
+    public static boolean isDistinctServiceReference(TypeSymbol typeSymbol) {
+        if (typeSymbol.typeKind() != TypeDescKind.TYPE_REFERENCE) {
+            return false;
+        }
+        TypeReferenceTypeSymbol typeReferenceTypeSymbol = (TypeReferenceTypeSymbol) typeSymbol;
+        Symbol typeDescriptor = typeReferenceTypeSymbol.typeDescriptor();
+        if (typeDescriptor.kind() != SymbolKind.CLASS) {
+            return false;
+        }
+        ClassSymbol classSymbol = (ClassSymbol) typeDescriptor;
+        return classSymbol.qualifiers().containsAll(Arrays.asList(Qualifier.SERVICE, Qualifier.DISTINCT));
     }
 }
