+++ compiler-plugin/src/main/java/io/ballerina/stdlib/graphql/compiler/validator/ServiceAnalysisTask.java
@@ -1,0 +1,104 @@
+/*
+ * Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.ballerina.stdlib.graphql.compiler.validator;
+
+import io.ballerina.compiler.api.symbols.ServiceDeclarationSymbol;
+import io.ballerina.compiler.api.symbols.SymbolKind;
+import io.ballerina.compiler.api.symbols.TypeDescKind;
+import io.ballerina.compiler.api.symbols.TypeSymbol;
+import io.ballerina.compiler.api.symbols.UnionTypeSymbol;
+import io.ballerina.compiler.syntax.tree.ServiceDeclarationNode;
+import io.ballerina.projects.plugins.AnalysisTask;
+import io.ballerina.projects.plugins.SyntaxNodeAnalysisContext;
+import io.ballerina.stdlib.graphql.compiler.validator.errors.CompilationError;
+import io.ballerina.tools.diagnostics.Diagnostic;
+import io.ballerina.tools.diagnostics.DiagnosticSeverity;
+
+import java.util.List;
+
+import static io.ballerina.stdlib.graphql.compiler.Utils.isGraphqlModuleSymbol;
+import static io.ballerina.stdlib.graphql.compiler.validator.ValidatorUtils.updateContext;
+
+/**
+ * Validates a Ballerina GraphQL Service.
+ */
+public class ServiceAnalysisTask implements AnalysisTask<SyntaxNodeAnalysisContext> {
+    private final ServiceValidator serviceValidator;
+
+    public ServiceAnalysisTask() {
+        this.serviceValidator = new ServiceValidator();
+    }
+
+    @Override
+    public void perform(SyntaxNodeAnalysisContext context) {
+        List<Diagnostic> diagnostics = context.semanticModel().diagnostics();
+        for (Diagnostic diagnostic : diagnostics) {
+            if (diagnostic.diagnosticInfo().severity() == DiagnosticSeverity.ERROR) {
+                return;
+            }
+        }
+        if (!isGraphQlService(context)) {
+            return;
+        }
+        this.serviceValidator.initialize(context);
+        this.serviceValidator.validate();
+    }
+
+    private boolean isGraphQlService(SyntaxNodeAnalysisContext context) {
+        ServiceDeclarationNode node = (ServiceDeclarationNode) context.node();
+        if (context.semanticModel().symbol(node).isEmpty()) {
+            return false;
+        }
+        if (context.semanticModel().symbol(node).get().kind() != SymbolKind.SERVICE_DECLARATION) {
+            return false;
+        }
+        ServiceDeclarationSymbol symbol = (ServiceDeclarationSymbol) context.semanticModel().symbol(node).get();
+        if (!hasGraphqlListener(symbol)) {
+            return false;
+        }
+        if (symbol.listenerTypes().size() > 1) {
+            updateContext(context, CompilationError.INVALID_MULTIPLE_LISTENERS, node.location());
+            return false;
+        }
+        return true;
+    }
+
+    private boolean hasGraphqlListener(ServiceDeclarationSymbol symbol) {
+        for (TypeSymbol listener : symbol.listenerTypes()) {
+            if (isGraphqlListener(listener)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    private boolean isGraphqlListener(TypeSymbol typeSymbol) {
+        if (typeSymbol.typeKind() == TypeDescKind.UNION) {
+            UnionTypeSymbol unionTypeSymbol = (UnionTypeSymbol) typeSymbol;
+            for (TypeSymbol member : unionTypeSymbol.memberTypeDescriptors()) {
+                if (isGraphqlModuleSymbol(member)) {
+                    return true;
+                }
+            }
+        } else {
+            return isGraphqlModuleSymbol(typeSymbol);
+        }
+        return false;
+    }
+}
